Add option to do a reverse rsync in the mellanox-lustre vagrantfile
Vagrant provides the ability to rsynce from the host to the guest vm but
it does not provide the ability to rsynce from the guest vm back to the
host. This is unfortunate because we need to download the files that are
generated. To solve this issue, give the user the option to run "vagrant
rsynce --reverse".

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>